+++ src/test/java/org/jabref/logic/bibtex/LatexFieldFormatterTests.java
@@ -18,7 +18,7 @@ public class LatexFieldFormatterTests {
     }
 
     @Test
-    public void normalizeNewlineInAbstractField() {
+    public void normalizeNewlineInAbstractField() throws Exception {
         String fieldName = "abstract";
         String text = "lorem" + OS.NEWLINE + " ipsum lorem ipsum\nlorem ipsum \rlorem ipsum\r\ntest";
 
@@ -34,7 +34,7 @@ public class LatexFieldFormatterTests {
     }
 
     @Test
-    public void preserveNewlineInAbstractField() {
+    public void preserveNewlineInAbstractField() throws Exception {
         String fieldName = "abstract";
         // The newlines are normalized according to the globally configured newline setting in the formatter
         String text = "lorem ipsum lorem ipsum" + OS.NEWLINE + "lorem ipsum lorem ipsum" + OS.NEWLINE;
@@ -46,7 +46,7 @@ public class LatexFieldFormatterTests {
     }
 
     @Test
-    public void preserveMultipleNewlinesInAbstractField() {
+    public void preserveMultipleNewlinesInAbstractField() throws Exception {
         String fieldName = "abstract";
         // The newlines are normalized according to the globally configured newline setting in the formatter
         String text = "lorem ipsum lorem ipsum" + OS.NEWLINE + OS.NEWLINE + "lorem ipsum lorem ipsum"
@@ -59,7 +59,7 @@ public class LatexFieldFormatterTests {
     }
 
     @Test
-    public void preserveNewlineInReviewField() {
+    public void preserveNewlineInReviewField() throws Exception {
         String fieldName = "review";
         // The newlines are normalized according to the globally configured newline setting in the formatter
         String text = "lorem ipsum lorem ipsum" + OS.NEWLINE + "lorem ipsum lorem ipsum" + OS.NEWLINE;
@@ -82,29 +82,29 @@ public class LatexFieldFormatterTests {
         assertEquals(expected, any);
     }
 
-    @Test(expected = IllegalArgumentException.class)
-    public void reportUnbalancedBracing() {
+    @Test(expected = InvalidFieldValueException.class)
+    public void reportUnbalancedBracing() throws Exception {
         String unbalanced = "{";
 
         formatter.format(unbalanced, "anyfield");
     }
 
-    @Test(expected = IllegalArgumentException.class)
-    public void reportUnbalancedBracingWithEscapedBraces() {
+    @Test(expected = InvalidFieldValueException.class)
+    public void reportUnbalancedBracingWithEscapedBraces() throws Exception {
         String unbalanced = "{\\}";
 
         formatter.format(unbalanced, "anyfield");
     }
 
     @Test
-    public void tolerateBalancedBrace() {
+    public void tolerateBalancedBrace() throws Exception {
         String text = "Incorporating evolutionary {Measures into Conservation Prioritization}";
 
         assertEquals("{" + text + "}", formatter.format(text, "anyfield"));
     }
 
     @Test
-    public void tolerateEscapeCharacters() {
+    public void tolerateEscapeCharacters() throws Exception {
         String text = "Incorporating {\\O}evolutionary {Measures into Conservation Prioritization}";
 
         assertEquals("{" + text + "}", formatter.format(text, "anyfield"));